+++ frontend/src/entities/User/model/selectors/getUserLoading/getUserLoading.ts
@@ -1,3 +1,3 @@
 import { StateSchema } from '../../../../../app/providers/StoreProvider/config/StateSchema';
 
-export const getUserIsLoading = (state: StateSchema) => state.user.isLoading;
+export const getUserLoading = (state: StateSchema) => state.user.loading;
