+++ frontend/src/pages/Home/ui/Home/home.scss
@@ -1,4 +1,4 @@
-@import '../../app/styles/mixins/global.scss';
+@import '../../../../app/styles/mixins/global.scss';
 
 .home {
   @include header-footer;
@@ -109,4 +109,3 @@
     }
   }
 }
-
