+++ frontend/src/pages/Home/ui/ItemsPreview/ItemsPreview.module.scss
