+++ frontend/src/pages/Home/ui/ItemsPreview/ItemsPreview.tsx
