+++ frontend/src/pages/Home/ui/ItemsPreviewSkeleton/ItemsPreviewSkeleton.module.scss
