+++ frontend/src/pages/Home/ui/ItemsPreviewSkeleton/ItemsPreviewSkeleton.tsx
