+++ frontend/src/pages/OrdersAdmin/index.ts
@@ -1,0 +1,3 @@
+import OrdersAdmin from './ui/OrdersAdmin/OrdersAdmin';
+
+export default OrdersAdmin;
