+++ frontend/src/pages/OrdersAdmin/ui/ChangeStatus/changeStatus.scss
@@ -1,4 +1,4 @@
-@import '../../../app/styles/mixins/global.scss';
+@import '../../../../app/styles/mixins/global.scss';
 
 .changeStatus {
   grid-column: 1/3;
