+++ frontend/src/pages/OrdersAdmin/ui/OrderDetails/orderDetails.scss
@@ -1,4 +1,4 @@
-@import '../../../app/styles/mixins/global.scss';
+@import '../../../../app/styles/mixins/global.scss';
 
 .orderDetails {
   display: grid;
