+++ frontend/src/pages/OrdersAdmin/ui/OrdersAdmin/ordersAdmin.scss
@@ -1,4 +1,4 @@
-@import '../../app/styles/mixins/global.scss';
+@import '../../../../app/styles/mixins/global.scss';
 
 .orders {
   @include main-block;
