--- frontend/src/widgets/ChangeStatus/index.ts
@@ -1,3 +1,0 @@
-import ChangeStatus from './ui/ChangeStatus';
-
-export { ChangeStatus };
