--- frontend/src/widgets/ItemsPreview/index.ts
@@ -1,3 +1,0 @@
-import ItemsPreview from './ui/ItemsPreview/ItemsPreview';
-
-export { ItemsPreview };
