--- frontend/src/widgets/OrderDetails/index.ts
@@ -1,2 +1,0 @@
-import OrderDetails from './ui/OrderDetails';
-export { OrderDetails };
