--- frontend/src/widgets/OrdersAdminListItem/index.ts
@@ -1,2 +1,0 @@
-import OrdersAdminListItem from './ui/OrdersAdminListItem';
-export { OrdersAdminListItem };
